put task into separate package and renamed runner class
