+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/two/PurchaseService.java
@@ -1,4 +1,4 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.two;
 
 public interface PurchaseService {
     boolean performPurchase(User buyer, User seller, Product product);
