+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/two/UserInformationService.java
@@ -1,4 +1,4 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.two;
 
 public interface UserInformationService {
     void informUser(User user, String message);
