+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/two/WebPurchaseService.java
@@ -1,4 +1,4 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.two;
 
 public class WebPurchaseService implements PurchaseService {
     @Override
